Fail if the first action is unauthenticated
Instead of checking if any of the actions are allowed by any of the
plugins, merely inspect the first action -- the process of completing it
may alter the rights on the rest of the actions.  This also makes both
authentication failure, and specifying the reason for the failure, much
easier, as individual CheckACL plugins can simply call MailError and
FAILURE with a custom message.